+++ files/paths_windows.go
@@ -1,5 +1,4 @@
 //go:build windows
-// +build windows
 
 package files
 
@@ -9,14 +8,16 @@ import (
 	"os"
 	"syscall"
 
-	"github.com/spf13/afero"
+	"github.com/go-git/go-billy/v5"
 )
 
-func checkFileMode(fs afero.Fs, name, filename string, expected os.FileMode, strict bool) error {
+func checkFileMode(fs billy.Filesystem, name, filename string, expected os.FileMode, strict bool) error {
 	var syserr syscall.Errno
 
-	if err := fs.Chmod(filename, expected); err != nil && strict && (!errors.As(err, &syserr) || syserr != 0) {
-		return fmt.Errorf("setting permissions for %s: %w", name, syserr)
+	if chfs, ok := fs.(billy.Change); ok {
+		if err := chfs.Chmod(filename, expected); err != nil && strict && (!errors.As(err, &syserr) || syserr != 0) {
+			return fmt.Errorf("setting permissions for %s: %w", name, syserr)
+		}
 	}
 
 	return nil