+++ sdk/git/exchange_test.go
@@ -2,25 +2,26 @@ package git_test
 
 import (
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 	"testing"
 
+	"github.com/go-git/go-billy/v5/memfs"
 	"github.com/julian7/redact/logger"
 	"github.com/julian7/redact/sdk/git"
 	"github.com/julian7/tester"
-	"github.com/spf13/afero"
 )
 
 func genGitRepo() (*git.Repo, error) {
 	r := &git.Repo{
-		Fs:       afero.NewMemMapFs(),
-		Logger:   logger.New(),
-		Common:   ".git",
-		Toplevel: "/git/repo",
+		Filesystem: memfs.New(),
+		Logger:     logger.New(),
+		Common:     ".git",
+		Toplevel:   "/git/repo",
 	}
 
-	err := r.Mkdir(r.Keydir(), 0700)
+	err := r.Filesystem.MkdirAll(r.Keydir(), 0700)
 	if err != nil {
 		return nil, fmt.Errorf("creating key dir %s: %w", r.Keydir(), err)
 	}
@@ -38,8 +39,8 @@ func writeKey(r *git.Repo) error {
 }
 
 func writeKX(r *git.Repo) error {
-	kxdir := filepath.Join(r.Toplevel, ".redact")
-	if err := r.MkdirAll(kxdir, 0755); err != nil {
+	kxdir := "/.redact"
+	if err := r.Filesystem.MkdirAll(kxdir, 0755); err != nil {
 		return fmt.Errorf("creating exchange dir: %w", err)
 	}
 
@@ -52,12 +53,12 @@ func writeKX(r *git.Repo) error {
 }
 
 func writeFile(r *git.Repo, fname string, perms os.FileMode, contents string) error {
-	of, err := r.OpenFile(fname, os.O_CREATE|os.O_WRONLY, perms)
+	of, err := r.Filesystem.OpenFile(fname, os.O_CREATE|os.O_WRONLY, perms)
 	if err != nil {
 		return fmt.Errorf("creating %s file: %w", fname, err)
 	}
 
-	if _, err := of.WriteString(contents); err != nil {
+	if _, err := io.WriteString(of, contents); err != nil {
 		return fmt.Errorf("writing %s file: %w", fname, err)
 	}